Extract CompareDist from examples to fe.atom_mapping
* Rename from CompareDist to CompareDistNonterminal
